# web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mosic 2.0 API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/tags": {
            "get": {
                "tags": [
                    "Etichette"
                ],
                "summary": "Lista etichette",
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo",
                        "examples": {
                            "application/json": {
                                "response": 200,
                                "total_results": 22,
                                "limit": "99999",
                                "offset": 0,
                                "data": [
                                    {
                                        "id": 2,
                                        "denominazione": "secondo tags"
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            },
            "post": {
                "tags": [
                    "Etichette"
                ],
                "summary": "Creazione etichetta",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo"
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            }
        },
        "/api/tags/{id}": {
            "get": {
                "tags": [
                    "Etichette"
                ],
                "summary": "Singola etichetta",
                "operationId": "idTag",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id dell'etichetta",
                        "required": true,
                        "type": "integer",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo",
                        "examples": {
                            "application/json": {
                                "response": 200,
                                "total_results": 22,
                                "limit": "99999",
                                "offset": 0,
                                "data": [
                                    {
                                        "id": 2,
                                        "denominazione": "secondo tags"
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            },
            "put": {
                "tags": [
                    "Etichette"
                ],
                "summary": "Salvataggio etichetta",
                "operationId": "idEtichetta",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id dell'etichetta",
                        "required": true,
                        "type": "integer",
                        "items": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "etichette",
                        "in": "body",
                        "description": "Richiesta",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "denominazione": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo",
                        "examples": {
                            "application/json": {
                                "id": 1,
                                "codice": 0,
                                "denominazione": "Documenti di seduta",
                                "descrizione": "Telex, Appunto generale, passi, etc...",
                                "id_uffici": 2
                            }
                        }
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Etichette"
                ],
                "summary": "Eliminazione etichetta",
                "operationId": "idEtichetta",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id dell'etichetta",
                        "required": true,
                        "type": "integer",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo"
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    },
                    "409": {
                        "description": "Etichetta \u00e8 associata ad un registro o ad un fascicolo o ad una delibera, impossibile eliminarla."
                    }
                }
            }
        },
...
    },
...
}